Move `IsPSPDisabled` helper function to `v1beta1/helper`
